move get_state to ff16
